fix:(backend):get agent list
* fix:(backend):get agent list



* feat:(backend): optimize shell



* feat:update



* refactor:(backend): optimize shell



See merge request: !224

+++ backend/application/singleagent/single_agent.go
@@ -22,7 +22,9 @@ import (
 	"code.byted.org/flow/opencoze/backend/domain/plugin/consts"
 	pluginEntity "code.byted.org/flow/opencoze/backend/domain/plugin/entity"
 	"code.byted.org/flow/opencoze/backend/domain/plugin/service"
+	searchEntity "code.byted.org/flow/opencoze/backend/domain/search/entity"
 	workflowEntity "code.byted.org/flow/opencoze/backend/domain/workflow/entity"
+
 	"code.byted.org/flow/opencoze/backend/pkg/errorx"
 	"code.byted.org/flow/opencoze/backend/pkg/lang/conv"
 	"code.byted.org/flow/opencoze/backend/pkg/lang/ptr"
@@ -37,8 +39,8 @@ var SingleAgentSVC = SingleAgentApplicationService{}
 
 func (s *SingleAgentApplicationService) UpdateSingleAgentDraft(ctx context.Context, req *playground.UpdateDraftBotInfoAgwRequest) (*playground.UpdateDraftBotInfoAgwResponse, error) {
 	// TODO： 这个一上来就查询？ 要做个简单鉴权吧？
-	botID := req.BotInfo.GetBotId()
-	currentAgentInfo, err := singleAgentDomainSVC.GetSingleAgent(ctx, botID, "")
+	agentID := req.BotInfo.GetBotId()
+	currentAgentInfo, err := singleAgentDomainSVC.GetSingleAgent(ctx, agentID, "")
 	if err != nil {
 		return nil, err
 	}
@@ -47,7 +49,7 @@ func (s *SingleAgentApplicationService) UpdateSingleAgentDraft(ctx context.Conte
 		return nil, errorx.New(errno.ErrPermissionCode, errorx.KV("msg", "bot_id invalidate"))
 	}
 
-	allow, err := permissionDomainSVC.CheckSingleAgentOperatePermission(ctx, botID, currentAgentInfo.SpaceID)
+	allow, err := permissionDomainSVC.CheckSingleAgentOperatePermission(ctx, agentID, currentAgentInfo.SpaceID)
 	if err != nil {
 		return nil, err
 	}
@@ -69,8 +71,8 @@ func (s *SingleAgentApplicationService) UpdateSingleAgentDraft(ctx context.Conte
 	}
 
 	if req.BotInfo.VariableList != nil {
-		botID = req.BotInfo.GetBotId()
-		varsMetaID, err := s.upsertVariableList(ctx, botID, *userID, "", req.BotInfo.VariableList)
+		agentID = req.BotInfo.GetBotId()
+		varsMetaID, err := s.upsertVariableList(ctx, agentID, *userID, "", req.BotInfo.VariableList)
 		if err != nil {
 			return nil, err
 		}
@@ -79,6 +81,21 @@ func (s *SingleAgentApplicationService) UpdateSingleAgentDraft(ctx context.Conte
 	}
 
 	err = singleAgentDomainSVC.UpdateSingleAgentDraft(ctx, agentInfo)
+	if err != nil {
+		return nil, err
+	}
+
+	err = domainNotifier.PublishApps(ctx, &searchEntity.AppDomainEvent{
+		DomainName: searchEntity.SingleAgent,
+		OpType:     searchEntity.Updated,
+		Agent: &searchEntity.Agent{
+			ID: agentID,
+			// SpaceID:      spaceID,
+			OwnerID:      *userID,
+			Name:         req.BotInfo.GetName(),
+			HasPublished: false,
+		},
+	})
 	if err != nil {
 		return nil, err
 	}
@@ -192,6 +209,21 @@ func (s *SingleAgentApplicationService) CreateSingleAgentDraft(ctx context.Conte
 	}
 
 	agentID, err := singleAgentDomainSVC.CreateSingleAgentDraft(ctx, userId, do)
+	if err != nil {
+		return nil, err
+	}
+
+	err = domainNotifier.PublishApps(ctx, &searchEntity.AppDomainEvent{
+		DomainName: searchEntity.SingleAgent,
+		OpType:     searchEntity.Created,
+		Agent: &searchEntity.Agent{
+			ID:           agentID,
+			SpaceID:      spaceID,
+			OwnerID:      userId,
+			Name:         do.Name,
+			HasPublished: false,
+		},
+	})
 	if err != nil {
 		return nil, err
 	}
@@ -313,7 +345,26 @@ func (s *SingleAgentApplicationService) GetDraftBotInfo(ctx context.Context, req
 }
 
 func (s *SingleAgentApplicationService) DeleteAgentDraft(ctx context.Context, req *developer_api.DeleteDraftBotRequest) (*developer_api.DeleteDraftBotResponse, error) {
+	uid := ctxutil.GetUIDFromCtx(ctx)
+	if uid == nil {
+		return nil, errorx.New(errno.ErrPermissionCode, errorx.KV("msg", "session required"))
+	}
+
 	err := singleAgentDomainSVC.DeleteAgentDraft(ctx, req.GetSpaceID(), req.GetBotID())
+	if err != nil {
+		return nil, err
+	}
+
+	err = domainNotifier.PublishApps(ctx, &searchEntity.AppDomainEvent{
+		DomainName: searchEntity.SingleAgent,
+		OpType:     searchEntity.Created,
+		Agent: &searchEntity.Agent{
+			ID:           req.GetBotID(),
+			SpaceID:      req.GetSpaceID(),
+			OwnerID:      *uid,
+			HasPublished: false,
+		},
+	})
 	if err != nil {
 		return nil, err
 	}
